♻️ (FTabs): Rename FTabContainer -> FTabsContainer
