+++ package.json
@@ -98,9 +98,9 @@
       "require": "./dist/FModal/useFModalDialog.cjs.js",
       "import": "./dist/FModal/useFModalDialog.es.js"
     },
-    "./FTabs/FTabContainer": {
-      "require": "./dist/FTabs/FTabContainer.cjs.js",
-      "import": "./dist/FTabs/FTabContainer.es.js"
+    "./FTabs/FTabsContainer": {
+      "require": "./dist/FTabs/FTabsContainer.cjs.js",
+      "import": "./dist/FTabs/FTabsContainer.es.js"
     },
     "./FTabs/FTab": {
       "require": "./dist/FTabs/FTab.cjs.js",
