+++ src/FTabs/FTab.mdx
@@ -6,9 +6,9 @@ import * as FTabStories from './FTab.stories';
 
 # FTab
 
-Registers a tab and its content inside a `FTabContainer`. Renders slot content only when the tab is active.
+Registers a tab and its content inside a `FTabsContainer`. Renders slot content only when the tab is active.
 
-> Must be used inside `FTabContainer`, otherwise `inject` will return `undefined` and throw an error.
+> Must be used inside `FTabsContainer`, otherwise `inject` will return `undefined` and throw an error.
 
 <Canvas of={FTabStories.Default} />
 <Controls of={FTabStories.Default} />
